+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature (`@grafana/faro-web-sdk`): track `web vitals` attribution (#595).
 - Feature (`@grafana/faro-web-sdk`): set span context for navigation events (#608).
+- Feature (`@grafana/faro-react`): add helper functions to initialize React Router integration (#622).
 
 - Enhancement (`@grafana/faro-web-sdk`): Auto extend a session if the Faro receiver indicates that a
   session is invalid (#591).